rearranging networking code
