+++ pydantic/__init__.py
@@ -6,6 +6,7 @@ from .error_wrappers import ValidationError
 from .errors import *
 from .fields import Required
 from .main import BaseConfig, BaseModel, Extra, compiled, create_model, validate_model
+from .networks import *
 from .parse import Protocol
 from .schema import Schema
 from .types import *
