+++ pydantic/types.py
@@ -1,41 +1,17 @@
 import json
 import re
 from decimal import Decimal
-from ipaddress import (
-    IPv4Address,
-    IPv4Interface,
-    IPv4Network,
-    IPv6Address,
-    IPv6Interface,
-    IPv6Network,
-    _BaseAddress,
-    _BaseNetwork,
-)
 from pathlib import Path
 from types import new_class
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    Dict,
-    Generator,
-    List,
-    Optional,
-    Pattern,
-    Set,
-    Tuple,
-    Type,
-    TypeVar,
-    Union,
-    cast,
-    no_type_check,
-)
+from typing import TYPE_CHECKING, Any, Callable, Dict, Generator, List, Optional, Pattern, Type, TypeVar, Union, cast
 from uuid import UUID
 
 from . import errors
-from .utils import AnyType, change_exception, host_tld_regex, import_string, url_regex, validate_email
+from .utils import AnyType, import_string
 from .validators import (
     bytes_validator,
+    constr_length_validator,
+    constr_strip_whitespace,
     decimal_validator,
     float_validator,
     int_validator,
@@ -64,12 +40,6 @@ __all__ = [
     'conlist',
     'ConstrainedStr',
     'constr',
-    'EmailStr',
-    'AnyUrl',
-    'AnyHttpUrl',
-    'HttpUrl',
-    'urlstr',
-    'NameEmail',
     'PyObject',
     'ConstrainedInt',
     'conint',
@@ -89,9 +59,6 @@ __all__ = [
     'DirectoryPath',
     'Json',
     'JsonWrapper',
-    'IPvAnyAddress',
-    'IPvAnyInterface',
-    'IPvAnyNetwork',
     'SecretStr',
     'SecretBytes',
     'StrictBool',
@@ -104,7 +71,6 @@ NoneStrBytes = Optional[StrBytes]
 OptionalInt = Optional[int]
 OptionalIntFloat = Union[OptionalInt, float]
 OptionalIntFloatDecimal = Union[OptionalIntFloat, Decimal]
-NetworkType = Union[str, bytes, int, Tuple[Union[str, bytes, int], Union[str, int]]]
 
 if TYPE_CHECKING:  # pragma: no cover
     from .fields import Field
@@ -248,173 +214,6 @@ class StrictBool(int):
             return value
 
         raise errors.StrictBoolError()
-
-
-class AnyUrl(str):
-    strip_whitespace = True
-    min_length = 1
-    max_length = 2 ** 16
-    allowed_schemes: Optional[Set[str]] = None
-    tld_required: bool = False
-    user_required: bool = False
-
-    __slots__ = ('scheme', 'user', 'password', 'host', 'port', 'path', 'query', 'fragment')
-
-    @no_type_check
-    def __new__(cls, url: str, **kwargs: Optional[str]) -> object:
-        return str.__new__(cls, url)
-
-    def __init__(
-        self,
-        url: str,
-        *,
-        scheme: str,
-        user: Optional[str] = None,
-        password: Optional[str] = None,
-        host: str,
-        port: Optional[str] = None,
-        path: Optional[str] = None,
-        query: Optional[str] = None,
-        fragment: Optional[str] = None,
-    ) -> None:
-        str.__init__(url)
-        self.scheme = scheme
-        self.user = user
-        self.password = password
-        self.host = host
-        self.port = port
-        self.path = path
-        self.query = query
-        self.fragment = fragment
-
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        yield not_none_validator
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: Any, field: 'Field', config: 'BaseConfig') -> 'AnyUrl':
-        if type(value) == cls:
-            return value
-        value = str_validator(value)
-        if cls.strip_whitespace:
-            value = value.strip()
-        url: str = cast(str, constr_length_validator(value, field, config))
-
-        m = url_regex.match(url)
-        if not m:  # pragma: no cover
-            # FIXME can this actually ever happen?
-            raise errors.UrlError()
-
-        parts = m.groupdict()
-        # debug(parts)
-        scheme = parts['scheme']
-        if scheme is None:
-            raise errors.UrlSchemeError()
-        if cls.allowed_schemes and scheme.lower() not in cls.allowed_schemes:
-            raise errors.UrlSchemePermittedError(cls.allowed_schemes)
-
-        if cls.user_required and parts['user'] is None:
-            raise errors.UrlUserInfoError()
-
-        host = parts['host']
-        if host is None:
-            raise errors.UrlHostError()
-        elif cls.tld_required and not host_tld_regex.fullmatch(host):
-            raise errors.UrlHostTldError()
-
-        if m.end() != len(url):
-            raise errors.UrlExtraError(extra=url[m.end() :])
-
-        return cls(url, **parts)
-
-    @no_type_check
-    def strip(self) -> str:
-        # required so constr_length_validator doesn't simplify AnyStr objects to strings
-        return self
-
-    def __repr__(self) -> str:
-        extra = ' '.join(f'{n}={getattr(self, n)!r}' for n in self.__slots__ if getattr(self, n) is not None)
-        return f'<{type(self).__name__}({super().__repr__()} {extra})>'
-
-
-class AnyHttpUrl(AnyUrl):
-    allowed_schemes = {'http', 'https'}
-
-
-class HttpUrl(AnyUrl):
-    allowed_schemes = {'http', 'https'}
-    tld_required = True
-
-
-class PostgresDsn(AnyUrl):
-    allowed_schemes = {'postgres', 'postgresql'}
-    user_required = True
-
-
-class RedisDsn(AnyUrl):
-    allowed_schemes = {'redis'}
-    user_required = True
-
-
-def urlstr(
-    *,
-    strip_whitespace: bool = True,
-    min_length: int = 1,
-    max_length: int = 2 ** 16,
-    tld_required: bool = True,
-    allowed_schemes: Optional[Set[str]] = None,
-) -> Type[str]:
-    # use kwargs then define conf in a dict to aid with IDE type hinting
-    namespace = dict(
-        strip_whitespace=strip_whitespace,
-        min_length=min_length,
-        max_length=max_length,
-        tld_required=tld_required,
-        allowed_schemes=allowed_schemes,
-    )
-    return type('UrlStrValue', (AnyUrl,), namespace)
-
-
-class EmailStr(str):
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        # included here and below so the error happens straight away
-        if email_validator is None:
-            raise ImportError('email-validator is not installed, run `pip install pydantic[email]`')
-
-        yield str_validator
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: str) -> str:
-        return validate_email(value)[1]
-
-
-class NameEmail:
-    __slots__ = 'name', 'email'
-
-    def __init__(self, name: str, email: str):
-        self.name = name
-        self.email = email
-
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        if email_validator is None:
-            raise ImportError('email-validator is not installed, run `pip install pydantic[email]`')
-
-        yield str_validator
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: str) -> 'NameEmail':
-        return cls(*validate_email(value))
-
-    def __str__(self) -> str:
-        return f'{self.name} <{self.email}>'
-
-    def __repr__(self) -> str:
-        return f'<NameEmail("{self}")>'
 
 
 class PyObject:
@@ -653,56 +452,6 @@ class Json(metaclass=JsonMeta):
             raise errors.JsonTypeError()
 
 
-class IPvAnyAddress(_BaseAddress):
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: Union[str, bytes, int]) -> Union[IPv4Address, IPv6Address]:
-        try:
-            return IPv4Address(value)
-        except ValueError:
-            pass
-
-        with change_exception(errors.IPvAnyAddressError, ValueError):
-            return IPv6Address(value)
-
-
-class IPvAnyInterface(_BaseAddress):
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: NetworkType) -> Union[IPv4Interface, IPv6Interface]:
-        try:
-            return IPv4Interface(value)
-        except ValueError:
-            pass
-
-        with change_exception(errors.IPvAnyInterfaceError, ValueError):
-            return IPv6Interface(value)
-
-
-class IPvAnyNetwork(_BaseNetwork):  # type: ignore
-    @classmethod
-    def __get_validators__(cls) -> 'CallableGenerator':
-        yield cls.validate
-
-    @classmethod
-    def validate(cls, value: NetworkType) -> Union[IPv4Network, IPv6Network]:
-        # Assume IP Network is defined with a default value for ``strict`` argument.
-        # Define your own class if you want to specify network address check strictness.
-        try:
-            return IPv4Network(value)
-        except ValueError:
-            pass
-
-        with change_exception(errors.IPvAnyNetworkError, ValueError):
-            return IPv6Network(value)
-
-
 class SecretStr:
     @classmethod
     def __get_validators__(cls) -> 'CallableGenerator':
@@ -753,25 +502,3 @@ class SecretBytes:
 
     def get_secret_value(self) -> bytes:
         return self._secret_value
-
-
-def constr_length_validator(v: 'StrBytes', field: 'Field', config: 'BaseConfig') -> 'StrBytes':
-    v_len = len(v)
-
-    min_length = field.type_.min_length or config.min_anystr_length  # type: ignore
-    if min_length is not None and v_len < min_length:
-        raise errors.AnyStrMinLengthError(limit_value=min_length)
-
-    max_length = field.type_.max_length or config.max_anystr_length  # type: ignore
-    if max_length is not None and v_len > max_length:
-        raise errors.AnyStrMaxLengthError(limit_value=max_length)
-
-    return v
-
-
-def constr_strip_whitespace(v: 'StrBytes', field: 'Field', config: 'BaseConfig') -> 'StrBytes':
-    strip_whitespace = field.type_.strip_whitespace or config.anystr_strip_whitespace  # type: ignore
-    if strip_whitespace:
-        v = v.strip()
-
-    return v
